updates StringArray to use TEXT primitive
The current implementation of StringArray uses VARCHAR and seems undesirable for any realworld use. Furter more, the postgresql docs use TEXT in its examples, and I feel dm-pg-types should follow suit.

This change re-implements StringArray to use TEXT and drops the length option entirely. If this is too heavy handed, let me know, and I will re-submit with a new type, TextArray, so that the old (and, IMHO, useless) implementation can be left as is.